+++ src/devices/mod.rs
@@ -1,3 +1,4 @@
 pub mod monitor_lem1802;
+pub mod clock_generic;
 pub mod keyboard_generic;
 pub mod floppy_m35fd;